+++ eObrazovanje/src/main/java/ftn/tseo/eEducation/service/ExamService.java
@@ -22,7 +22,7 @@ public class ExamService {
 		return examRepository.save(exam);
 	}
 
-	public void remove(Integer id) {
+	public void remove(Long id) {
 		examRepository.deleteById(id);
 	}
